install: Check that open symlink exist before trying to create it

+++ install.sh
@@ -269,7 +269,7 @@ done
 if ./base/.usr/bin/_onlinux ; then
     # Make a symlink from open to xdg-open, so that `open` or the `o` alias work the same on Linux
     # or macOS
-    ln -s /usr/bin/xdg-open ~/.usr/bin/open
+    [[ -L ~/.usr/bin/open ]] || ln -s /usr/bin/xdg-open ~/.usr/bin/open
 
     # Symlink Linux specific dotfiles
     for dir in *-linux ; do
